rename search components for clarity
